+++ .cmake-format.yaml
@@ -1,3 +1,4 @@
+---
 parse:
   additional_commands:
     catkin_package:
@@ -9,20 +10,22 @@ parse:
         LIBRARIES: '*'
     catkin_add_gtest:
       pargs: 1
+    pybind11_add_module:
+      pargs: 1
 format:
   line_width: 88
   tab_size: 2
   use_tabchars: false
-  max_subgroups_hwrap: 5
-  #max_pargs_hwrap: 6
+  max_subgroups_hwrap: 3
+  max_pargs_hwrap: 2
   max_rows_cmdline: 2
   dangle_parens: true
   dangle_align: prefix
   min_prefix_chars: 4
   max_prefix_chars: 10
-  max_lines_hwrap: 2
+  max_lines_hwrap: 1
   command_case: canonical
   keyword_case: unchanged
-  always_wrap: [add_library]
+  always_wrap: [add_library, pybind11_add_module]
   enable_sort: true
   autosort: false
