+++ .github/workflows/cmake.yaml
@@ -11,7 +11,7 @@ jobs:
       # setup package dependencies
       - name: Dependencies
         run: |
-          sudo apt install libgtest-dev libeigen3-dev nlohmann-json3-dev libzmq3-dev python3-dev
+          sudo apt install libgtest-dev libeigen3-dev nlohmann-json3-dev libzmq3-dev pybind11-dev python3-dev
       # configure repo
       - name: Configure
         run: |
