+++ .github/workflows/python.yaml
@@ -20,7 +20,7 @@ jobs:
       - name: Dependencies
         run: |
           python -m pip install --upgrade pip
-          sudo apt install libgtest-dev libeigen3-dev nlohmann-json3-dev libzmq3-dev python3-dev
+          sudo apt install libgtest-dev libeigen3-dev nlohmann-json3-dev libzmq3-dev pybind11-dev python3-dev
       # build and install code
       - name: Install
         run: python -m pip install ${{github.workspace}}
