+++ package.xml
@@ -15,6 +15,7 @@
   <depend>eigen</depend>
   <depend>libzmqpp-dev</depend>
   <depend>nlohmann-json-dev</depend>
+  <depend>pybind11-dev</depend>
   <depend>python3-dev</depend>
 
   <export>
